requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
asgiref==3.8.1
    # via
    #   -r requirements.in
    #   django
astroid==3.3.8
    # via
    #   -r requirements.in
    #   pylint
atomicwrites==1.4.1
    # via -r requirements.in
beautifulsoup4==4.12.3
    # via -r requirements.in
boto3==1.36.2
    # via django-ses
botocore==1.36.2
    # via
    #   boto3
    #   s3transfer
bsdiff4==1.2.5
    # via -r requirements.in
certifi==2024.12.14
    # via requests
chardet==5.2.0
    # via django-newsletter
charset-normalizer==3.4.1
    # via requests
concurrent-log-handler==0.9.25
    # via -r requirements.in
dill==0.3.9
    # via
    #   -r requirements.in
    #   pylint
django==5.1.5
    # via
    #   -r requirements.in
    #   django-braces
    #   django-classy-tags
    #   django-cookie-law
    #   django-newsletter
    #   django-ses
    #   django-tinymce
    #   django-user-agents
django-admin==2.0.2
    # via -r requirements.in
django-braces==1.16.0
    # via -r requirements.in
django-classy-tags==4.1.0
    # via
    #   -r requirements.in
    #   django-cookie-law
django-cookie-law==2.2.0
    # via -r requirements.in
django-environ==0.12.0
    # via -r requirements.in
django-excel-response2==3.0.6
    # via django-admin
django-newsletter==1.0
    # via -r requirements.in
django-ses==4.3.2
    # via -r requirements.in
django-sitetree==1.18.0
    # via -r requirements.in
django-six==1.0.5
    # via
    #   django-admin
    #   django-excel-response2
django-tinymce==4.1.0
    # via -r requirements.in
django-user-agents==0.4.0
    # via -r requirements.in
excel-base==1.0.4
    # via django-excel-response2
factory-boy==3.3.1
    # via -r requirements.in
faker==33.3.1
    # via factory-boy
feedparser==6.0.11
    # via -r requirements.in
feedreader==0.3.1
    # via -r requirements.in
future==1.0.0
    # via -r requirements.in
idna==3.10
    # via requests
iniconfig==2.0.0
    # via
    #   -r requirements.in
    #   pytest
isort==5.13.2
    # via
    #   -r requirements.in
    #   pylint
isoweek==1.3.3
    # via timeconvert
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
lazy-object-proxy==1.10.0
    # via -r requirements.in
ldif3==3.1.1
    # via
    #   -r requirements.in
    #   django-newsletter
lxml==5.3.0
    # via
    #   -r requirements.in
    #   feedreader
mccabe==0.7.0
    # via
    #   -r requirements.in
    #   pylint
more-itertools==10.6.0
    # via -r requirements.in
packaging==24.2
    # via pytest
pillow==11.1.0
    # via django-newsletter
platformdirs==4.3.6
    # via
    #   -r requirements.in
    #   pylint
pluggy==1.5.0
    # via
    #   -r requirements.in
    #   pytest
portalocker
    # via
    #   -r requirements.in
    #   concurrent-log-handler
protobuf==5.29.3
    # via -r requirements.in
pylint==3.3.3
    # via -r requirements.in
pyrfc3339==2.0.1
    # via -r requirements.in
pytest==8.3.4
    # via -r requirements.in
pytest-runner==6.0.1
    # via -r requirements.in
python-card-me==0.9.3
    # via
    #   -r requirements.in
    #   django-newsletter
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   faker
    #   feedreader
    #   python-card-me
    #   timeconvert
redis==5.2.1
    # via -r requirements.in
requests==2.32.3
    # via requests-mock
requests-mock==1.12.1
    # via -r requirements.in
s3transfer==0.11.1
    # via boto3
screen==1.0.1
    # via excel-base
sgmllib3k==1.0.0
    # via
    #   -r requirements.in
    #   feedparser
six==1.17.0
    # via python-dateutil
sorl-thumbnail==12.11.0
    # via
    #   -r requirements.in
    #   sorl-watermark
sorl-watermark==1.2.0
    # via -r requirements.in
soupsieve==2.6
    # via
    #   -r requirements.in
    #   beautifulsoup4
sqlparse==0.5.3
    # via
    #   -r requirements.in
    #   django
surlex==0.2.0
    # via -r requirements.in
text-unidecode==1.3
    # via -r requirements.in
timeconvert==3.0.13
    # via excel-base
tomlkit==0.13.2
    # via
    #   -r requirements.in
    #   pylint
typing-extensions==4.12.2
    # via faker
tzlocal==5.2
    # via timeconvert
ua-parser==1.0.0
    # via
    #   -r requirements.in
    #   user-agents
ua-parser-builtins==0.18.0.post1
    # via ua-parser
unicodecsv==0.14.1
    # via
    #   -r requirements.in
    #   django-newsletter
urllib3==2.3.0
    # via
    #   botocore
    #   requests
user-agents==2.2.0
    # via
    #   -r requirements.in
    #   django-user-agents
wcwidth==0.2.13
    # via -r requirements.in
wrapt==1.17.2
    # via -r requirements.in
xlwt==1.3.0
    # via excel-base
django_azure_communication_email
django-simple-captcha
setuptools
django.debug_toolbar